Enrollment/src/main/java/com/example/Enrollment/MemberController.java
package com.example.Enrollment;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.BindingResult;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import jakarta.validation.Valid;

@RestController
public class MemberController {

    private final MemberService memberService;

    @Autowired
    public MemberController(MemberService memberService) {
        this.memberService = memberService;
    }
   

    @PostMapping("/api/members/enroll")
    public String enrollMember( @Valid MemberDTO memberDTO, BindingResult bindingResult) {
        if (bindingResult.hasErrors()) {
            return "error";
        }

        boolean isEnrollmentSuccessful = memberService.enrollMember(memberDTO);
        if (isEnrollmentSuccessful) {
            return "Member enrollment successful";
        } else {
            return "Failed to enroll member";
        }
    }
}
